+++ R/sandbox.R
@@ -214,17 +214,23 @@ renv_sandbox_task <- function(...) {
   if (!renv_sandbox_activated())
     return()
 
+  # allow opt-out if necessary
   enabled <- getOption("renv.sandbox.task", default = TRUE)
   if (!enabled)
     return()
 
-  # make sure the sandbox exists
+  # get sandbox path
   sandbox <- tail(.libPaths(), n = 1L)
+
+  # make sure it exists
   if (!file.exists(sandbox)) {
     warning("the renv sandbox was deleted; it will be re-generated", call. = FALSE)
     ensure_directory(sandbox)
     renv_sandbox_generate(sandbox)
   }
+
+  # update the sandbox write time / mtime
+  Sys.setFileTime(sandbox, time = Sys.time())
 
 }